+++ example/lib/core/failures/api_server_failure.dart
@@ -3,6 +3,4 @@ import 'failure.dart';
 class ApiServerFailure extends Failure {
   final String message;
   const ApiServerFailure(this.message);
-  @override
-  List<Object?> get props => [];
 }