+++ example/lib/core/failures/api_timeout_failure.dart
@@ -1,6 +1,3 @@
 import 'failure.dart';
 
-class ApiTimeOutFailure extends Failure {
-  @override
-  List<Object?> get props => [];
-}
+class ApiTimeOutFailure extends Failure {}