+++ example/lib/core/failures/exception_failure.dart
@@ -3,6 +3,4 @@ import 'failure.dart';
 class ExceptionFailure extends Failure {
   final Exception message;
   const ExceptionFailure(this.message);
-  @override
-  List<Object?> get props => [message];
 }