+++ example/lib/core/failures/file_empty_failure.dart
@@ -1,7 +1,3 @@
 import 'failure.dart';
 
-class FileEmptyFailure extends Failure {
-  @override
-  // TODO: implement props
-  List<Object?> get props => throw UnimplementedError();
-}
+class FileEmptyFailure extends Failure {}