+++ example/lib/core/failures/login_keycloak_failure.dart
@@ -1,6 +1,3 @@
 import 'failure.dart';
 
-class LoginKeyCloakFailure extends Failure {
-  @override
-  List<Object?> get props => [];
-}
+class LoginKeyCloakFailure extends Failure {}