+++ example/lib/core/use_case.dart
@@ -3,6 +3,5 @@ import 'package:dartz/dartz.dart';
 import 'failures/failure.dart';
 
 abstract class UseCase<Type, Params> {
-  @override
   Future<Either<Failure, Type>> call(Params params);
 }